Refactored Privates
Added functionality to add private lessons to the table and database via the AddRow in the PrivatesTable.
Added functionality to remove private lessons from the table and database via the RemoveRows in the PrivatesTable.
Added functionality to update input fields in the Privates table.
Added the ID in the RemoveRow instances to the parameters passed back to the updateCallback (utilized in both the Instructors and Privates components).
Added functionality to lookup instructor IDs by their name in the Privates component.
Added the SectionDescription and Tutorial specializations.
Fixed an error where multiple private lessons would not be attributed to the same instructor.
Remove the seconds from the Privates time entry in the database when formatting the database response in the Connector.
Remove the deprecated methods that use jQuery from the Privates component.
Updated Input and TableCell utils to accept and display numbers as props.