Display release version
In the Header and Footer components, display information of the current release of the app.
Re-structure the Footer so that it only displays version data.
Instead of passing the Connector component to the InstructorPreferences and Privates components as a prop, pass only the function required by the component.
Remove the styleHeader prop because it was always a no-op.