Set up FnScroll specification
Since all the test cases are bundled together by webpack before being funnelled to AVA, 't.context' is shared by all files. To remedy, each file will have their own sub-section of the test context.

This will be a problem in the future, and can maybe be remedied by compiling the app and test imports with webpack and running the tests as-written.